Buscar chegadas atrasadas por data finalizado

+++ api/check_if/src/main/java/br/edu/ifsul/sapucaia/check_if/domain/Professor.java
@@ -25,7 +25,7 @@ public class Professor {
 
     private String email;
 
-    private int celular;
+    private Long celular;
 
     private boolean notificacaoWhatsapp;
